requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
beautifulsoup4==4.14.2
lxml==6.0.2
requests==2.32.5
feedparser==6.0.12
python-dotenv==1.2.1
